fix(ingestionbench): stop a mistyped flag printing every credential
The re-review after the rebase found a real one, plus three smaller faults and
the file-size rule.

SECRETS. `-database-url`, `-clickhouse` and `-tenants` took their defaults from
the environment, so the resolved Postgres URL, the ClickHouse URL and every
tenant's API key were each a flag's DefValue. `flag.ContinueOnError` still runs
the usage function when parsing fails, and that prints every default — so ONE
mistyped flag would have written all of them to stderr, which the workflow
captures and uploads as an artifact. The flags now default to empty and resolve
from the environment after parsing, so the explicit flag still wins and nothing
secret is ever a default.

COUNTS. `numberToInt` swallowed a decode error and returned 0. Those values feed
the violation rules directly, so a count ClickHouse rendered in a shape the
driver could not read became "the map projection dropped N spans" for data
sitting in the table the whole time. countRow now uses clickhouseCount, which
already fails on an unreadable value, so the run says "could not run" instead of
"you lost data".

TENANTS. Two identical entries passed the two-tenant floor and then defeated the
check that floor exists for: the cross-tenant query would have compared a
project against itself, found nothing foreign, and reported isolation as
PASSING. Duplicate project ids are refused.

CANCELLATION. The burst loop kept dispatching windows after the context was
done. Every one of those requests fails at the socket and is counted as a
Failure, so a cancelled run would have reported a large RequestsFailed figure
the pipeline never caused.

Also: the metrics-server manifest is pinned instead of tracking `releases/latest`
(it is applied with cluster-admin, and that URL can change image and permissions
with no review here); `span_event_type` is on the dispatch form, since the docs
promise every driver knob is and this is the one value the PR flags as
unconfirmed; and driver.go is split along the seams it already had, because the
repo fails a new file over 300 source lines and it had reached 1328. Every file
in the package is now under it.

Two of my own tests were wrong and are fixed: one asserted `stageSeed(99, X)`
equals itself, which holds for any implementation including a broken one, and
the ordering test pinned only the non-scattering branch, so deleting the scatter
call would still have passed.
